refactor: move types
